# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Travel Platform Dashboard API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardOverview {
  bronzePois: number;
  silverPois: number;
  goldPois: number;
  cities: number;
  quarantined: number;
  pipelineRuns: number;
  dataLineageEdges: number;
  avgQualityScore: number;
  avgRating: number;
}

export interface CityPoiCount {
  city: string;
  cityName: string;
  count: number;
}

export interface CategoryPoiCount {
  category: string;
  count: number;
}

export interface PipelineFunnel {
  bronze: number;
  silver: number;
  gold: number;
  quarantine: number;
}

export interface QualityBucket {
  range: string;
  count: number;
}

export interface RatingBucket {
  range: string;
  count: number;
}

export interface PipelineExecution {
  executionId: string;
  pipelineName: string;
  status: string;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  cities: string[];
  categories: string[];
  recordsProcessed: number;
  recordsFailed: number;
  currentStage?: string;
}

export interface SyncStateSummary {
  city: string;
  category: string;
  currentCount: number;
  status: string;
  upsertedCount: number;
}

export interface City {
  cityCode: string;
  name: string;
  nameEn: string;
  region: string;
  population: number;
  /** @nullable */
  description?: string | null;
}

export interface Poi {
  poiId: string;
  name: string;
  city: string;
  cityName: string;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  reviewCount: number | null;
  /** @nullable */
  qualityScore: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  priceLevel?: number | null;
}

export interface PoisResponse {
  pois: Poi[];
  total: number;
}

export interface QuarantineReason {
  rule: string;
  count: number;
}

export type GetPoisParams = {
city?: string;
category?: string;
limit?: number;
offset?: number;
};

export type GetTopRatedPoisParams = {
city?: string;
category?: string;
limit?: number;
};