# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Travel Platform Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pois:
    get:
      operationId: getPois
      tags: [pois]
      summary: Browse gold-layer POIs
      description: Returns paginated gold-layer POIs with optional filters
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: POIs list with total count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoisResponse"
  /pois/top-rated:
    get:
      operationId: getTopRatedPois
      tags: [pois]
      summary: Top rated POIs
      description: Returns top-rated gold-layer POIs
      parameters:
        - name: city
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Top rated POIs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Poi"
...
components:
  schemas:
...
    Poi:
      type: object
      required: [poiId, name, city, cityName, category, rating, reviewCount, qualityScore]
      properties:
        poiId:
          type: string
        name:
          type: string
        city:
          type: string
        cityName:
          type: string
        category:
          type: string
        subcategory:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
        reviewCount:
          type: ["integer", "null"]
        qualityScore:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        website:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        priceLevel:
          type: ["integer", "null"]
    PoisResponse:
      type: object
      required: [pois, total]
      properties:
        pois:
          type: array
          items:
            $ref: "#/components/schemas/Poi"
        total:
          type: integer
...