# lib/api-zod/src/generated/types/poi.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Travel Platform Dashboard API
 * OpenAPI spec version: 0.1.0
 */

export interface Poi {
  poiId: string;
  name: string;
  city: string;
  cityName: string;
  category: string;
  /** @nullable */
  subcategory?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  rating: number | null;
  /** @nullable */
  reviewCount: number | null;
  /** @nullable */
  qualityScore: number | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  priceLevel?: number | null;
}