# lib/api-zod/src/generated/types/poisResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smart Travel Platform Dashboard API
 * OpenAPI spec version: 0.1.0
 */
import type { Poi } from './poi';

export interface PoisResponse {
  pois: Poi[];
  total: number;
}